Remove test_scheduling
Test scheduling is basically a feature to analyze the scheduling_info
struct, which was a bad idea, as said in previous commit message. It
was also used to analyze timings for loading configuration, which there
is another API for already (there are lot of timing_points in the code,
use naemon -T to display).

So this is basically a cleanup to remove redundant timing analysis
features, which just adds lot of noise to the code.

Signed-off-by: Max Sikstrom <msikstrom@op5.com>